fix websocket port not available on heroku
