--- server/ws/index.html
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8" />
-        <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <title>Document</title>
-    </head>
-    <body></body>
-
-    <script>
-        const url = "ws://localhost:4040";
-        const connection = new WebSocket(url);
-
-        connection.onopen = () => {
-            connection.send("connection initiated");
-        };
-
-        connection.onerror = (error) => {
-            console.log(`WebSocket error: ${error}`);
-        };
-    </script>
-</html>
